Harden native execution and pin accelerator equivalence
Make native MEX dispatch prove its equivalence to the Python oracle and fall back transactionally for operations whose integer, floating-point, NaN, reduction, or exception behavior is not bit-exact. Correct tile bounds, signed arithmetic, FP accumulation and signed-zero handling, scalar aperture routing, constructor-state synchronization, and native-prefix fallback accounting.

Pin every borrowed Python memory export with validated capacity and guest-range metadata. Serialize execution, attachment replacement, and framebuffer conversion without recursive shared-mutex use or GIL lock cycles; release replaced exports after unlocking, zero incomplete frames, and prevent the display from bypassing native scanout bounds through raw-memory fallback.

Add assertive dispatch/oracle, buffer-lifetime, callback-reentry, renderer-ordering, wrapper-exception, and display regressions. The focused suite passes 226 tests sequentially, legacy MEX/tile checks pass, and a separate ASAN plus UBSAN build passes the 230-test focused and legacy set.

+++ accel/mp64_crypto.h
@@ -1020,7 +1020,7 @@ struct WotsChain {
 
     // Memory pointer (set by CPUState init, points to main RAM)
     uint8_t* mem;
-    uint32_t mem_size;
+    uint64_t mem_size;
 
     // SHA3 engine pointer (shared with CryptoDevices)
     CryptoSHA3* sha3;
